pyproject.toml
[tool.poetry]
name = "tpiuo"
version = "0.1.0"
description = ""
authors = ["Marko Selendic <Marko.Selendic@fer.hr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
azure-eventhub = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
